退出会议

+++ src/main/java/com/yang/service/UserService.java
@@ -59,17 +59,19 @@ public class UserService {
         }
     }
 
-    public void quitMeeting(String title, String userName) {
+    public boolean quitMeeting(String title, String userName) {
         User user = userDao.findByName(userName);
         Agenda agenda = agendaDao.findByTitle(title);
-        if (user.isParticipator()) {
+        if (user.isParticipator()&&agendaDao.isParticipator(agenda, userName)!=null) {
             user.setParticipator(false);
             userDao.addUser(user);
-            agendaDao.deleteParticipator(agenda, userName);
+            agendaDao.deleteParticipator(agenda,user);
             agendaDao.addAgenda(agenda);
             if (agenda.getParticipator().isEmpty()) {
                 agendaDao.deleteAgenda(title);
             }
+        return true;
         }
+        return false;
     }
 }
